projects/telemetry/datavisualization/telemetry/requirements.txt
cantools==37.2.0
fakeredis==2.10.3
pyserial==3.5
python_can==4.0.0
redis==3.5.3
redistimeseries==1.4.5
